change package ID to make F-Droid happy use library from maven fix minor bugs: - make text in sync error message activity readable (was white on grey...) - repair the preferences link in android settings - don't crash when clicking "done" on already created account
