+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="de.danielweisser.android.ldapsync"
-    android:versionCode="14"
-    android:versionName="2.1" android:installLocation="internalOnly">
+    package="de.wikilab.android.ldapsync"
+    android:versionCode="15"
+    android:versionName="2.1.1" android:installLocation="internalOnly">
 
 
     <uses-permission android:name="android.permission.GET_ACCOUNTS" />
@@ -23,7 +23,7 @@
         android:label="@string/app_name"
         android:theme="@style/LightThemeSelector" >
         <service
-            android:name=".authenticator.LDAPAuthenticationService"
+            android:name="de.wikilab.android.ldapsync.authenticator.LDAPAuthenticationService"
             android:exported="true"
             android:permission="android.permission.MANAGE_ACCOUNTS" >
             <intent-filter>
@@ -35,7 +35,7 @@
                 android:resource="@xml/authenticator" />
         </service>
         <service
-            android:name=".syncadapter.SyncService"
+            android:name="de.wikilab.android.ldapsync.syncadapter.SyncService"
             android:exported="true"
             android:permission="android.permission.MANAGE_ACCOUNTS" >
             <intent-filter>
@@ -48,7 +48,7 @@
         </service>
 
         <activity
-            android:name=".authenticator.LDAPAuthenticatorActivity"
+            android:name="de.wikilab.android.ldapsync.authenticator.LDAPAuthenticatorActivity"
             android:excludeFromRecents="true"
             android:label="@string/ui_activity_title"
             android:windowSoftInputMode="stateVisible|adjustResize">
@@ -60,10 +60,10 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".authenticator.AccountSettingsActivity" >
+        <activity android:name="de.wikilab.android.ldapsync.authenticator.AccountSettingsActivity" >
         </activity>
 
-        <activity android:name=".authenticator.AccountsListActivity" >
+        <activity android:name="de.wikilab.android.ldapsync.authenticator.AccountsListActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -85,7 +85,7 @@
             </intent-filter>
         </activity>
         -->
-        <activity android:name=".activity.SyncErrorActivity"
+        <activity android:name="de.wikilab.android.ldapsync.activity.SyncErrorActivity"
             android:label="UF Phonebook Sync Error"
             android:launchMode="singleTask"
             android:taskAffinity=""
