--- app/src/main/java/de/danielweisser/android/ldapsync/Util.java
@@ -1,40 +1,0 @@
-package de.danielweisser.android.ldapsync;
-
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.io.Writer;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-
-public class Util {
-	public static SharedPreferences getPrefs(Context ctx) {
-		return PreferenceManager.getDefaultSharedPreferences(ctx.getApplicationContext());
-	}
-
-	public static String getStackTrace(Throwable aThrowable) {
-		final Writer result = new StringWriter();
-		final PrintWriter printWriter = new PrintWriter(result);
-		aThrowable.printStackTrace(printWriter);
-		return result.toString();
-	}
-
-	/**
-	 * Defines a custom format for the stack trace as String.
-	 */
-	public static String getCustomStackTrace(Throwable aThrowable) {
-		// add the class name and any message passed to constructor
-		final StringBuilder result = new StringBuilder();
-		result.append(aThrowable.toString());
-		final String NEW_LINE = System.getProperty("line.separator");
-		result.append(NEW_LINE);
-
-		// add each element of the stack trace
-		for (StackTraceElement element : aThrowable.getStackTrace()) {
-			result.append(element);
-			result.append(NEW_LINE);
-		}
-		return result.toString();
-	}
-}
