--- app/src/main/java/de/danielweisser/android/ldapsync/client/ContactTest.java
@@ -1,22 +1,0 @@
-package de.danielweisser.android.ldapsync.client;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class ContactTest {
-	
-	private Contact c = new Contact();
-
-	@Test
-	public void testDn() {
-		c.setDn("testDn");
-		assertEquals("testDn", c.getDn());
-	}
-
-	@Test
-	public void testGetFirstName() {
-		c.setFirstName("testFirstName");
-		assertEquals("testFirstName", c.getFirstName());
-	}
-}
