+++ app/src/main/java/de/wikilab/android/ldapsync/Constants.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2010 Daniel Weisser
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.wikilab.android.ldapsync;
+
+/**
+ * Constants for the LDAP sync adapter.
+ * 
+ * @author <a href="mailto:daniel.weisser@gmx.de">Daniel Weisser</a>
+ */
+public class Constants {
+
+	/**
+	 * Account type string.
+	 */
+	public static final String ACCOUNT_TYPE = "de.wikilab.android.ldapsync";
+
+	/**
+	 * SD card LDAPSync folder.
+	 */
+	public static final String SDCARD_FOLDER = "/LDAPSync";
+	
+	public static final String PARAM_USERNAME = "username";
+	public static final String PARAM_PASSWORD = "password";
+	public static final String PARAM_HOST = "host";
+	public static final String PARAM_PORT = "port";
+	public static final String PARAM_ENCRYPTION = "encryption";
+
+}
