+++ app/src/main/res/layout/sync_error_layout.xml
@@ -9,7 +9,6 @@
     android:layout_width="fill_parent" 
     android:layout_height="wrap_content" 
     android:padding="10dp"
-    android:textColor="#FFFFFF"
     />
 <Button
  android:id="@+id/buttonSendErrorReport"
