+++ app/src/main/res/xml/authenticator.xml
@@ -16,7 +16,7 @@
  -->
  
 <account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-    android:accountType="de.danielweisser.android.ldapsync"
+    android:accountType="de.wikilab.android.ldapsync"
     android:icon="@drawable/icon"
     android:smallIcon="@drawable/icon"
     android:label="@string/label"
