+++ app/src/main/res/xml/syncadapter.xml
@@ -17,6 +17,6 @@
 
 <sync-adapter xmlns:android="http://schemas.android.com/apk/res/android"
     android:contentAuthority="com.android.contacts"
-    android:accountType="de.danielweisser.android.ldapsync"
+    android:accountType="de.wikilab.android.ldapsync"
     android:supportsUploading="false"
 />
